NodeIterator will handle the generation of N and send that value together with the table when extracted from the ExecutionNodes

+++ SQLDataProducer.Entities/ExecutionEntities/ExecutionNode.cs
@@ -133,8 +133,8 @@ namespace SQLDataProducer.Entities.ExecutionEntities
             return node;
         }
 
-        private int _repeatCount  = 1;
-        public int RepeatCount
+        private long _repeatCount  = 1;
+        public long RepeatCount
         {
             get
             {
